+++ src/main/java/de/dm/intellij/liferay/language/velocity/LiferayVtlMacroProvider.java
@@ -42,6 +42,9 @@ public class LiferayVtlMacroProvider extends VtlGlobalMacroProvider implements T
         } else if
                 (
                     (liferayVersion == LiferayVersions.LIFERAY_VERSION_7_0) ||
+                    (liferayVersion == LiferayVersions.LIFERAY_VERSION_7_1) ||
+                    (liferayVersion == LiferayVersions.LIFERAY_VERSION_7_2) ||
+                    (liferayVersion == LiferayVersions.LIFERAY_VERSION_7_3) ||
                     (liferayVersion == LiferayVersions.LIFERAY_VERSION_UNKNOWN)
                 ) {
             return "/com/liferay/vtl/VM_liferay_70.vm";